// src/app/shared/components/theme-toggle/enum/type.enum.ts
export enum ETypeThem {
  ThemeIndigo = 'Indigo',
  ThemePink = 'Pink',
  ThemeGreen = 'Green',
  ThemePurple = 'Purple',
  ThemeOrange = 'Orange',
  ThemeBrown = 'Brown',
  ThemeGrey = 'Grey',
  ThemeBlueGray = 'Blue Grey',
  ThemeDeepOrange = 'Deep Orange',
  ThemeTeal = 'Teal',
  ThemeLightGreen = 'Ligh Green',
  ThemeCyan = 'Cyan',
  ThemeLightBule = 'Light Blue',
  ThemeDeepPurple = 'Deep Purple',
  ThemeRed = 'Red',
  ThemeDeepPink = 'Deep Pink',
  ThemeLime = 'Lime',
}